Add OpenClaw one-click apply [P3.7]
